Various improvements aimed at getting local optimization regressions fixed: - Rename _gc to ##gc - Absolute labels are now supported - Generate _dispatch-label

+++ vm/code_block.cpp
@@ -159,7 +159,7 @@ cell compute_relocation(relocation_entry rel, cell index, code_block *compiled)
 	case RT_XT_PIC_TAIL:
 		return (cell)word_xt_pic_tail(untag<word>(ARG));
 	case RT_HERE:
-		return offset + (short)untag_fixnum(ARG);
+		return offset + untag_fixnum(ARG);
 	case RT_THIS:
 		return (cell)(compiled + 1);
 	case RT_STACK_CHAIN:
